Fixed potential performance issues by batching deletes into a single call and freeing select query memory

+++ system/ee/legacy/libraries/Session.php
@@ -431,28 +431,39 @@ class EE_Session {
 
 		if ((rand() % 100) < $this->gc_probability)
 		{
+			echo 'DELETED';
+			$sessions_to_delete = [];
 			$cp_expire = ee()->localize->now - $this->cpan_session_len;
 			$user_expire = ee()->localize->now - $this->user_session_len;
 
 			// Get only the CP sessions that have expired.
-			$admin_results = ee()->db->where('admin_sess', 1)
+			$admin_results = ee()->db->select('session_id')
+				->where('admin_sess', 1)
 				->where('last_activity < ', $cp_expire)
 				->get('sessions');
 
 			// Loop through the old sessions and delete their session and security_hash data.
 			foreach ($admin_results->result() as $row) {
-				ee()->db->where('session_id', $row->session_id);
-				ee()->db->delete(array('security_hashes', 'sessions'));
+				$sessions_to_delete[] = $row->session_id;
 			}
 
 			// Get only the user sessions that have expired.
-			$user_results = ee()->db->where('admin_sess', 0)
+			$user_results = ee()->db->select('session_id')
+				->where('admin_sess', 0)
 				->where('last_activity < ', $user_expire)
 				->get('sessions');
 
 			// Loop through the old sessions and delete their session and security_hash data.
 			foreach ($user_results->result() as $row) {
-				ee()->db->where('session_id', $row->session_id);
+				$sessions_to_delete[] = $row->session_id;
+			}
+
+			// Free the memory from this query as it's at the start of our execution loop.
+			$admin_results->free_result();
+			$user_results->free_result();
+
+			if (!empty($sessions_to_delete)) {
+				ee()->db->where_in('session_id', $sessions_to_delete);
 				ee()->db->delete(array('security_hashes', 'sessions'));
 			}
 		}
